# src/main/java/com/example/gymcrm/dto/user/UserCreateDto.java
package com.example.gymcrm.dto.user;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;
import lombok.experimental.SuperBuilder;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
@SuperBuilder
public class UserCreateDto {

    @ToString.Exclude
    @Schema(description = "The first name of the user", example = "john")
    @NotNull(message = "First name must not be null")
    @Size(min = 1, max = 50, message = "First name must be between 1 and 50 characters")
    private String firstName;

    @ToString.Exclude
    @Schema(description = "The first name of the user", example = "doe")
    @NotNull(message = "Last name must not be null")
    @Size(min = 1, max = 50, message = "Last name must be between 1 and 50 characters")
    private String lastName;
}